WIP: working on C interface in XReqLib library

+++ src/lib/xreq.h
@@ -21,16 +21,16 @@ typedef struct XReq_Table XReq_Table;
 struct XReq_Error;
 typedef struct XReq_Error XReq_Error;
 
-typedef unsigned int  XReq_Status;
-typedef unsigned long XReq_Duration;
-typedef const char*   XReq_Cstr;
-typedef int           XReq_Bool;
-typedef XReq_Cstr*    XReq_Tags;
+typedef unsigned short XReq_Status;
+typedef unsigned long  XReq_Duration;
+typedef const char*    XReq_Cstr;
+typedef short          XReq_Bool;
+typedef XReq_Cstr*     XReq_Tags;
 
 void XReq_Time_Start (XReq_Duration*);
 void XReq_Time_Stop  (XReq_Duration*);
 
-void XReq_Format_Set_Num_Tests (XReq_Format*, int);
+void XReq_Format_Set_Num_Tests (XReq_Format*, long);
 void XReq_Format_Free          (XReq_Format*);
 
 void XReq_Format_Start_Tests   (XReq_Format*);
@@ -54,7 +54,7 @@ void XReq_Format_Stop_Outline         (XReq_Format*);
 void XReq_Format_Enter_Scenario       (XReq_Format*);
 void XReq_Format_Start_Scenario       (XReq_Format*);
 void XReq_Format_Put_Scenario         (XReq_Format*, XReq_Cstr, XReq_Cstr, XReq_Tags);
-void XReq_Format_Put_Scenario_Outline (XReq_Format*, int, XReq_Cstr, XReq_Cstr, XReq_Tags);
+void XReq_Format_Put_Scenario_Outline (XReq_Format*, long, XReq_Cstr, XReq_Cstr, XReq_Tags);
 void XReq_Format_Stop_Scenario        (XReq_Format*);
 
 void XReq_Format_Start_Step (XReq_Format*);
@@ -65,7 +65,7 @@ void XReq_Format_Stop_Step  (XReq_Format*);
 XReq_Conditional* XReq_Conditional_New  ();
 void      XReq_Conditional_Free         (XReq_Conditional*);
 XReq_Bool XReq_Conditional_Eval_Tags    (XReq_Conditional*, XReq_Tags);
-XReq_Bool XReq_Conditional_Eval_Position(XReq_Conditional*, XReq_Cstr, int);
+XReq_Bool XReq_Conditional_Eval_Position(XReq_Conditional*, XReq_Cstr, long);
 
 XReq_Report* XReq_Report_New   ();
 void XReq_Report_step_Free     (XReq_Report*);
@@ -74,8 +74,8 @@ void XReq_Report_step_pass     (XReq_Report*);
 void XReq_Report_step_fail     (XReq_Report*);
 void XReq_Report_scenario_pass (XReq_Report*);
 void XReq_Report_scenario_fail (XReq_Report*);
-void XReq_Report_num_steps_inc (XReq_Report*, int);
-int  XReq_Report_get_num_steps (XReq_Report*);
+void XReq_Report_num_steps_inc (XReq_Report*, long);
+long XReq_Report_get_num_steps (XReq_Report*);
 XReq_Bool XReq_Report_Status   (XReq_Report*);
